- nuke on LUP

+++ www/projects/nuke-2.tcl
@@ -75,7 +75,7 @@ with_transaction {
 		where object_id = :project_id
 	)"
     ns_log Notice "projects/nuke-2: im_forum_topics"
-    db_dml forum "delete from im_forum_topics where project_id = :project_id"
+    db_dml forum "delete from im_forum_topics where object_id = :project_id"
 
 
     # Timesheet
@@ -85,12 +85,13 @@ with_transaction {
 
     # Translation Quality
     ns_log Notice "projects/nuke-2: im_trans_quality_entries"
-    db_dml trans_quality "delete from im_trans_quality_entries where report_id in (
-	select report_id from from im_trans_quality_reports where task_id in (select task_id from im_trans_tasks where project_id = :project_id)
-    )"
-    ns_log Notice "projects/nuke-2: im_trans_quality_reports"
-    db_dml trans_quality "delete from im_trans_quality_reports where task_id in (select task_id from im_trans_tasks where project_id = :project_id)";
-
+    if {[db_table_exists im_trans_quality_reports]} {
+	db_dml trans_quality "delete from im_trans_quality_entries where report_id in (
+	    select report_id from im_trans_quality_reports where task_id in (select task_id from im_trans_tasks where project_id = :project_id)
+        )"
+	ns_log Notice "projects/nuke-2: im_trans_quality_reports"
+	db_dml trans_quality "delete from im_trans_quality_reports where task_id in (select task_id from im_trans_tasks where project_id = :project_id)";
+    }
 
     # Translation
     ns_log Notice "projects/nuke-2: im_task_actions"
@@ -104,6 +105,7 @@ with_transaction {
     ns_log Notice "projects/nuke-2: im_trans_tasks"
     db_dml trans_tasks "delete from im_trans_tasks where project_id = :project_id"
 
+    db_dml project_target_languages "delete from im_target_languages where project_id = :project_id"
 
     # Filestorage
     ns_log Notice "projects/nuke-2: im_fs_folder_status"
@@ -128,7 +130,9 @@ with_transaction {
 
 
     ns_log Notice "users/nuke2: Main tables"
-    db_dml delete_user "delete from im_projects where project_id = :project_id"
+
+    db_dml parent_projects "update im_projects set parent_id = null where parent_id = :project_id"
+    db_dml delete_projects "delete from im_projects where project_id = :project_id"
 
 } {
